Pin yup_oauth2 versions to a particular git commit.
The revision hasn't been released on crates.io yet, but it includes a
change that makes the error returned by GetToken to be Send+Sync which
we desire for our own error type. Once a version that includes that
change is pushed to crates.io we will move to that. I expect that to
happen long before we're ready to release anything for general
consumption.